fix: `contracts/agent_execution.py` | 45 | contracts 层反向依赖 execution 层：`from dayu.execution.runtime_config import AgentRunningConfigSnapshot, RunnerRunningConfigSnapshot`。违反架构硬约束"禁止反向依赖"。`RunnerRunningConfigSnapshot` 和 `AgentRunningConfigSnapshot` 应下沉到 contracts 层，或通过 Protocol 消除此依赖。
