+++ tests/README.md
@@ -138,7 +138,7 @@ python -m pytest tests --cov=dayu --cov-report=term --cov-branch
 - `test_scene_execution.py` 与配置加载相关测试还需要守住 CLI runner 已彻底禁用：`runner_type=cli` 应在模型加载或 Host 装配阶段显式失败，而不是继续被当成可用路径。
 - 直接构造 `ResolvedExecutionOptions`、`AcceptedToolConfigSpec` 或 `PreparedAgentTurnSnapshot` 的测试，工具配置应通过 `toolset_configs` 进入契约；不要再把 `doc_tool_limits` / `fins_tool_limits` / `web_tools_config` 当成跨层字段夹带到 Host 或 contract snapshot。
 - `test_scene_execution.py` 还要守住内部 `RunnerType` / `FallbackMode` 可提升为枚举承接，但配置快照和 JSON 边界仍保持原字符串值，不要把外部输入边界改造成枚举字面量协议。
-- execution 运行配置的测试真源现在拆成两层：`dayu.execution.options` 负责合并规则，`dayu.execution.runtime_config` 负责纯运行配置模型与快照转换。凡是直接构造 `ResolvedExecutionOptions` 的测试，应优先使用 `AgentRuntimeConfig`、`OpenAIRunnerRuntimeConfig`、`CliRunnerRuntimeConfig` 这组 execution 侧纯模型，不要再把 engine 实现类直接塞回 execution 层对象。
+- execution 运行配置的测试真源现在拆成三层：`dayu.contracts.runtime_config_snapshot` 负责跨层快照 TypedDict（`RunnerRunningConfigSnapshot` / `AgentRunningConfigSnapshot`）；`dayu.execution.options` 负责合并规则；`dayu.execution.runtime_config` 负责纯运行配置模型与快照转换函数。凡是直接构造 `ResolvedExecutionOptions` 的测试，应优先使用 `AgentRuntimeConfig`、`OpenAIRunnerRuntimeConfig`、`CliRunnerRuntimeConfig` 这组 execution 侧纯模型，不要再把 engine 实现类直接塞回 execution 层对象。
 - `test_host_executor.py`、`test_run_registry.py`、`test_host_admin_service.py` 与 `test_web_routes.py` 需要共同守住 Host timeout 取消链路。`request_cancel()` 只记录 `cancel_requested_at` / `cancel_requested_reason`，终态 `cancel_reason` 必须等 Host 在 run 生命周期边界统一收敛后才写入；管理视图与 API 需要同时透出这两层信息。
 - `test_host_executor.py` 与 `test_run_registry.py` 还要继续守住 orphan run 修复语义：startup cleanup 只能清理超过 grace period 的 dead-pid 活跃 run，不能误伤刚启动或仍在收尾的前台 direct operation；cleanup 必须把超期 run 写入独立的 `RunState.UNSETTLED` 吸收态（不再复用 `FAILED`），owner 自愈的判据必须是 `state == UNSETTLED && owner_pid == os.getpid()`——不得再依赖 `error_summary` 字符串来区分 orphan 与业务 FAILED；异常路径也不能再被重复 `fail_run()` 掩盖成状态机错误，Agent stream / sync / agent 三条路径必须统一经由 preserving wrapper 收口。
 - `test_cli_graceful_shutdown.py` 负责守住 CLI 优雅退出边界：`register_cli_shutdown_hook()` 在 `_prepare_cli_host_dependencies` 装配 Host 后一次性挂钩 SIGTERM / SIGHUP / atexit，进程收到信号或正常退出时必须调 `Host.shutdown_active_runs_for_owner()` 把本 pid 的活跃 run 主动收敛到 CANCELLED；hook 必须是幂等的、信号注册失败时必须静默降级不炸掉 CLI。
@@ -199,7 +199,7 @@ python -m pytest tests --cov=dayu --cov-report=term --cov-branch
 - 项目级 `tests/conftest.py` 这类 import 兼容夹具若需要预加载模块，也要先显式断言 `ModuleSpec/loader` 非空，再在 `ModuleType` 边界集中 `cast(Any, module)` 后补动态别名；不要把 `spec.loader` 的可空分支或动态属性赋值散落到正文。
 - `test_cli_interactive_coverage.py` 这类 CLI UI 测试里的 fake session 应直接实现 `ChatServiceProtocol` / `PromptServiceProtocol` 公开方法（如 `submit_turn`、`resume_pending_turn`、`list_resumable_pending_turns`、`submit`），不要继续只靠旧的 `stream/stream_turn` 形态再由测试把对象硬塞进 `interactive_ui`。
 - `test_sse_parser.py` 与 `test_sec_pipeline_download.py` 这类 Engine/Fins 装配测试，遇到 `ClientResponse`、`AsyncOpenAIRunnerRunningConfig`、`SecDownloader` 等生产签名时，优先在测试装配边界用 helper 做一次显式类型收窄；不要为了消除 pyright 报错去让测试 stub 继承不匹配的生产实现类，尤其不要把同步 stub 硬继承到异步 downloader 真源上。ticker 市场识别一律对齐 `dayu.fins.ticker_normalization.normalize_ticker` 真源，需要 stub 时在消费者模块路径上 `monkeypatch.setattr(module, "normalize_ticker", ...)` 返回自造 `NormalizedTicker`，不要重造识别类。
-- Engine 内部测试如果必须直接构造 `AgentRunningConfig` 或 `AsyncOpenAIRunnerRunningConfig`，应从各自定义模块导入，不要再依赖 `dayu.engine` 的聚合导出面；跨层 runtime config 真源仍以 `dayu.execution.runtime_config` 为准。
+- Engine 内部测试如果必须直接构造 `AgentRunningConfig` 或 `AsyncOpenAIRunnerRunningConfig`，应从各自定义模块导入，不要再依赖 `dayu.engine` 的聚合导出面；跨层 runtime config 快照真源在 `dayu.contracts.runtime_config_snapshot`，纯运行配置值对象与转换函数在 `dayu.execution.runtime_config`。
 - `tests/fins/test_bs_report_form_common.py` 这类 processor 单测里，`Source` dummy 应保持普通类做结构兼容；`XBRL`、`_TableBlock` 等实现类型应在测试装配边界集中 `cast` / helper 收窄，不要把 `SimpleNamespace` 直接写进处理器字段或方法签名。
 - `tests/engine/test_bs_processor.py` 这类基础 processor 搜索测试里，`SearchHit.section_ref/snippet` 等非必填字段也应通过 helper 安全读取；不要在基础回归里裸下标访问，否则很容易和 `search_utils` 一类文件重复产生同源 TypedDict 错误。
 - `tests/engine/test_bs_processor_coverage.py` 这类覆盖补丁测试里，不要复用同名测试函数；传给 `_safe_table_text()` 的 `table_tag` 也要先断言非空，再把 BeautifulSoup 的可选返回值传进内部 helper。
